Fix ~1000 recurring build warnings across 4 root causes
All zero-behavior-change cleanups, verified via full rebuild (0 errors)
and full test suite (4113 ran / 4009 passed / 104 skipped / 0 failed,
exact baseline match).

1. std::min/std::max called from device code (616 warnings, #20013-D/
   #20015-D): ~30 unguarded call sites across ~15 src/shared headers
   (bxdfs_layered.h, light_bounds.h, compact_light_bounds.h,
   bilinear_patch.h, direction_cone.h, float_bits.h, scalar_math.h,
   punctual_lights.h, cloud_medium.h, shapes.h, sampling_sphere.h,
   sampling_sphere_cone.h). Applied this codebase's own established
   `#if defined(__CUDACC__) fminf/fmaxf #else std::min/std::max #endif`
   pattern (already used correctly in ~half these same files) rather
   than inventing a new helper. Templated functions genuinely
   instantiated with both float and double (confirmed per call site,
   not assumed) use unsuffixed fmin/fmax instead of fminf/fmaxf, since
   the 'f' variants silently truncate a double call to float precision.

   Verification surfaced 2 closely related variants of the same nvcc
   diagnostic family also worth closing while rebuilding repeatedly:
   std::numeric_limits<T>::infinity()/min() (color_utils.h, direction_cone.h,
   mis_sampling.h, scalar_math.h) swapped to HUGE_VALF/HUGE_VAL/FLT_MIN,
   and std::isinf (float_bits.h, sampling_sphere.h) swapped to CUDA's
   unqualified isinf() overload - same guard pattern, different device-safe
   replacement per function.

2. Redundant CPU_GPU on defaulted constructors (156 warnings, #20012-D):
   nvcc infers host/device eligibility automatically for a special member
   function defaulted on its first declaration - explicit annotations are
   ignored. Dropped CPU_GPU from the 7 affected constructors (color_utils.h
   XYZ, cloud_medium.h CloudMajorantIterator, grid_medium.h
   RayMajorantSegment, compensated_float.h CompensatedSum, materials.h
   MeasuredMaterial, measured_bxdf.h MeasuredBxDF, reservoir_sampler.h
   WeightedReservoirSampler).

3. Internal math_functions.h include (139 warnings): swapped to the
   toolkit-recommended cuda_runtime.h in the 7 files that included it
   directly (bxdfs_base.h, fresnel.h, noise.h, normal_map.h,
   punctual_lights.h, sampling_helpers.h, volume_scattering.h).

4. double->float narrowing (42 warnings, MSVC C4244): pbrt_gpu_builder.h's
   rgbgrid-medium builder (sigma_s_r/g/b, Le_r/g/b, gridDensity - the last
   of these found during verification, not in the original audit) now
   uses explicit std::transform+static_cast instead of .assign()/.insert()
   from a double range into a float vector. Same pattern applied to
   pbrt_cpu_builder.h's identical single-value textureGamma narrowing
   (found during verification - matches the cast already used 3x on the
   GPU side for this same field).

Intentionally NOT fixed (out of scope, confirmed real but different):
nanovdb's own C4146 warnings (vendored third-party code, not ours to
patch), 2 single-occurrence linker warnings (not worth chasing), and
noise.h's NoisePerm __device__-variable-read-from-host warning (#20091-D -
a genuine but separate architectural question, not a simple guard).

+++ src/TheRestOfYourLife/pbrt_cpu_builder.h
@@ -118,7 +118,10 @@ inline shared_ptr<texture> checkerOrMixSlot(const std::string &filename, const d
 // MipMapOptions{} (gamma 2.2, Clamp, no invert) - unchanged.
 inline MipMapOptions imageMapOptionsFor(const pbrt_flatten::Material &m) {
 	MipMapOptions opts;
-	opts.gamma = m.textureGamma;
+	// static_cast, matching the identical GPU-side cast already used for
+	// this same double(pbrt)->float(MipMapOptions) field 3x in
+	// gpu/optix/pbrt_gpu_builder_materials.h.
+	opts.gamma = static_cast<float>(m.textureGamma);
 	opts.invert = m.textureInvert;
 	// m.textureWrapIndex is the single, already-validated resolution of
 	// m.textureWrap (Material::textureWrapIndex's own comment, pbrt_flatten.h)
